requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
itsdangerous==2.2.0

SQLAlchemy==2.0.32
alembic==1.13.2

passlib[bcrypt]==1.7.4
pyotp==2.9.0

pydantic==2.8.2
python-dotenv==1.0.1

starlette==0.38.4
filetype==1.2.0

bcrypt==4.0.1

PyInstaller==6.19.0 